🏷️ Refactor: 타입 분리
